Seperated TableView into another class
