+++ src/main/java/ui/NewGUI.java
@@ -4,9 +4,6 @@ import core.Duke;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.control.*;
-import javafx.scene.control.cell.CheckBoxTableCell;
-import javafx.scene.control.cell.PropertyValueFactory;
-import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.StackPane;
@@ -36,36 +33,17 @@ public class NewGUI extends VBox {
     private ImageView userImage;
     @FXML
     private ImageView systemImage;
-
-    @FXML
-    private TableView<Task> taskView;
-    @FXML
-    private TableColumn<Task, String> taskTypeCol;
-    @FXML
-    private TableColumn<Task, Boolean> statusCol;
-    @FXML
-    private TableColumn<Task, String> descriptionCol;
-    @FXML
-    private TableColumn<Task, LocalDateTime> dateDueCol;
-    DateTimeFormatter outputFormat = DateTimeFormatter.ofPattern("dd MMMM hhmm a");
-    @FXML
-    private ObservableList<Task> tasks;
+    private TaskView generateTable;
 
     public void setDuke(Duke d) {
         duke = d;
-        taskView.setItems(duke.getAllTasks());
+        generateTable = new TaskView(duke.getAllTasks());
+        tableArea.getChildren().addAll(generateTable.getTable());
     }
 
     @FXML
     public void initialize() {
-        System.out.println("fk");
         systemOutput.setText("hi my name is duke");
-        taskTypeCol.setCellValueFactory(new PropertyValueFactory<>("taskType"));
-        statusCol.setCellValueFactory(new PropertyValueFactory<>("isDone"));
-        statusCol.setCellFactory(tc -> new CheckBoxTableCell<>());
-        descriptionCol.setCellValueFactory(new PropertyValueFactory<>("description"));
-        dateDueCol.setCellValueFactory(new PropertyValueFactory<>("dateDue"));
-        dateDueCol.setCellFactory(new NewGUI.ColumnFormatter<>(outputFormat));
     }
 
 
@@ -75,37 +53,9 @@ public class NewGUI extends VBox {
         String response = duke.getResponse(input);
         systemOutput.setText(response);
         userInput.clear();
-        taskView.setItems(duke.getAllTasks());
+        generateTable.setTable(duke.getAllTasks());
 
     }
-
-    private class ColumnFormatter<S, T> implements Callback<TableColumn<S, T>, TableCell<S, T>> {
-
-        private final DateTimeFormatter format;
-
-        public ColumnFormatter(DateTimeFormatter format) {
-            super();
-            this.format = format;
-        }
-
-        @Override
-        public TableCell<S, T> call(TableColumn<S, T> arg0) {
-            return new TableCell<S, T>() {
-                @Override
-                protected void updateItem(T item, boolean empty) {
-                    super.updateItem(item, empty);
-                    if (item == null || empty) {
-                        setGraphic(null);
-                    } else {
-                        LocalDateTime ld = (LocalDateTime) item;
-                        String val = ld.format(format);
-                        setGraphic(new Label(val));
-                    }
-                }
-            };
-        }
-    }
-
 
 
 }
