+++ src/main/resources/view/TaskView.fxml
@@ -3,11 +3,11 @@
 <?import javafx.scene.control.TableColumn?>
 <?import javafx.scene.control.TableView?>
 
-<fx:root maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" type="TableView" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/11.0.1">
+<fx:root fx:id="taskView" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" type="TableView" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
   <columns>
-    <TableColumn fx:id="taskType" prefWidth="98.66675055027008" text="Tasktype" />
-      <TableColumn fx:id="status" prefWidth="99.33336639404297" text="Status" />
-    <TableColumn fx:id="description" prefWidth="258.66664123535156" text="Description" />
-      <TableColumn fx:id="DateDue" minWidth="0.0" prefWidth="141.33334350585938" text="DateDue" />
+    <TableColumn fx:id="taskTypeCol" prefWidth="98.66675055027008" text="Tasktype" />
+      <TableColumn fx:id="statusCol" prefWidth="99.33336639404297" text="Status" />
+    <TableColumn fx:id="descriptionCol" prefWidth="258.66664123535156" text="Description" />
+      <TableColumn fx:id="dateDueCol" minWidth="0.0" prefWidth="141.33334350585938" text="DateDue" />
   </columns>
 </fx:root>
