+++ tasklist.txt
@@ -1,11 +1,2 @@
 todo [0]deadline
-todo [0]
-todo [0]1
-todo [0]afadsfasdgfasgsdgasdgfsadfasdfadsf
-todo [0]asd
-todo [0]asfshjkhcdxghjkjfdsanfsgdafngdgagdfsfsssaS
-deadline [0]do this shit/by02/02/2019 0800
-todo [0]
-todo [0]
-deadline [0]fasdfasd/by03/02/2019 0900
-deadline [0]asdfa/by02/02/2019 0700
+deadline [1]test/by02/02/2020 0800
